lad: clean up functions
handle-lists is replaced by handle-nums which handles numbers instead of lists.
just like its predecessor, it does the same thing except it converts the numbers to
list of digits itself instead.

extract-digit is a bit cleaner because of skipping one set! call

ladd is simple now because handle-nums does a bit more